Redis 적용하기

+++ src/main/java/com/study/sns/controller/PostController.java
@@ -58,7 +58,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId, Authentication authentication) {
+    public Response<Long> likeCount(@PathVariable Integer postId, Authentication authentication) {
         postService.likeCount(postId);
         return Response.success(postService.likeCount(postId));
     }
